reorganise files and fix errors
